fix(components page & minimal refactoring elsewhere): add icons to compoents page and refactor
Refactor components page to include the icons available to this project; move components related to
the home page to that subfolder; update Hero Section, Button Component and Search Box to use Icon
component
