+++ src/pages/home/svgWave.jsx
